move commcare deploy code to new package
