+++ src/commcare_cloud/commands/deploy.py
@@ -13,6 +13,7 @@ from commcare_cloud.commands import shared_args
 from commcare_cloud.commands.ansible import ansible_playbook
 from commcare_cloud.commands.ansible.helpers import AnsibleContext
 from commcare_cloud.commands.command_base import Argument, CommandBase
+from commcare_cloud.commands.deploy.commcare import deploy_commcare
 from commcare_cloud.commands.terraform.aws import get_default_username
 from commcare_cloud.environment.main import get_environment
 from commcare_cloud.environment.paths import get_available_envs
@@ -65,21 +66,15 @@ class Deploy(CommandBase):
     def run(self, args, unknown_args):
         check_branch(args)
         environment = get_environment(args.env_name)
-        deploy_revs = self._confirm_deploy_revs(environment, args, quiet=args.quiet)
 
         deploy_component = args.component
         if deploy_component == None:
             deploy_component = 'both' if environment.meta_config.always_deploy_formplayer else 'commcare'
 
         if deploy_component in ['commcare', 'both']:
-            print(color_summary("You are about to deploy the following code:"))
-            with indent():
-                for name, rev in deploy_revs:
-                    print(color_summary("{}: {}".format(name, rev)))
-            if ask('Continue with deploy?', quiet=args.quiet):
-                if deploy_component != 'both':
-                    _warn_no_formplayer()
-                self.deploy_commcare(environment, deploy_revs, args, unknown_args)
+            if deploy_component != 'both':
+                _warn_no_formplayer()
+            deploy_commcare(environment, args, unknown_args)
         if deploy_component in ['formplayer', 'both']:
             if deploy_component != 'both':
                 if args.commcare_rev:
@@ -88,16 +83,6 @@ class Deploy(CommandBase):
                     print(color_warning('--set does not apply to a formplayer deploy and will be ignored'))
             self._announce_formplayer_deploy_start(environment)
             self.deploy_formplayer(environment, args, unknown_args)
-
-    def deploy_commcare(self, environment, deploy_revs, args, unknown_args):
-        fab_func_args = self.get_deploy_commcare_fab_func_args(args)
-        fab_settings = [args.fab_settings] if args.fab_settings else []
-        for name, rev in deploy_revs:
-            var = 'code_branch' if name == 'commcare' else '{}_code_branch'.format(name)
-            fab_settings.append('{}={}'.format(var, rev))
-        commcare_cloud(environment.name, 'fab', 'deploy_commcare{}'.format(fab_func_args),
-                       '--set', ','.join(fab_settings),
-                       branch=args.branch, *unknown_args)
 
     def deploy_formplayer(self, environment, args, unknown_args):
         def run_ansible_playbook_command():
@@ -124,50 +109,6 @@ class Deploy(CommandBase):
         )
         if rc != 0:
             return rc
-
-    @staticmethod
-    def get_deploy_commcare_fab_func_args(args):
-        fab_func_args = []
-
-        if args.quiet:
-            fab_func_args.append('confirm=no')
-        if args.resume:
-            fab_func_args.append('resume=yes')
-        if args.skip_record:
-            fab_func_args.append('skip_record=yes')
-
-        if fab_func_args:
-            return ':{}'.format(','.join(fab_func_args))
-        else:
-            return ''
-
-    @staticmethod
-    def _confirm_deploy_revs(environment, args, quiet=False):
-        default_branch = environment.fab_settings_config.default_branch
-        branches = [
-            ('commcare', 'commcare_rev', default_branch),
-        ]
-        for repo in environment.meta_config.git_repositories:
-            branches.append((repo.name, '{}_rev'.format(repo.name), repo.version))
-
-        diffs = []
-        actuals = []
-        for repo_name, arg_name, default in branches:
-            actual = getattr(args, arg_name, None)
-            actuals.append((repo_name, actual or default))
-            if actual and actual != default:
-                diffs.append("'{}' repo: {} != {}".format(repo_name, default, actual))
-
-        if diffs:
-            message = (
-                "Whoa there bud! You're deploying non-default. "
-                "\n{}\n"
-                "ARE YOU DOING SOMETHING EXCEPTIONAL THAT WARRANTS THIS?"
-            ).format('/n'.join(diffs))
-            if not ask(message, quiet=quiet):
-                exit(-1)
-
-        return actuals
 
     @staticmethod
     def _announce_formplayer_deploy_start(environment):
